--- a/common/src/network/async_socket.hpp
+++ b/common/src/network/async_socket.hpp
@@ -74,31 +74,7 @@ public:
   }
 
   template <typename MessageTypeOut>
-  void asyncWrite(const MessageTypeOut &msg) {
-    if (!mSocket.is_open()) {
-      spdlog::error("Failed to write to the socket: not opened");
-      return;
-    }
-    auto buffer = Serializer::serialize(msg);
-    boost::endian::little_int16_at bodySize = static_cast<MessageSize>(buffer.size());
-    auto dataPtr = std::make_shared<ByteBuffer>(sizeof(MessageSize) + buffer.size());
-
-    std::memcpy(dataPtr->data(), &bodySize, sizeof(bodySize));
-    std::memcpy(dataPtr->data() + sizeof(bodySize), buffer.data(), buffer.size());
-
-    if constexpr (std::is_same_v<Socket, TcpSocket>) {
-      boost::asio::async_write(
-          mSocket, boost::asio::buffer(dataPtr->data(), dataPtr->size()),
-          [this, dataPtr](BoostErrorRef ec, size_t size) { writeHandler(ec, size); });
-    } else if constexpr (std::is_same_v<Socket, UdpSocket>) {
-      mSocket.async_send_to(
-          boost::asio::buffer(dataPtr->data(), dataPtr->size()), mEndpoint,
-          [this, dataPtr](BoostErrorRef ec, size_t size) { writeHandler(ec, size); });
-    }
-  }
-
-  template <typename MessageTypeOut>
-  void asyncWrite(const std::vector<MessageTypeOut> &msgVec) {
+  void asyncWrite(Span<MessageTypeOut> msgVec) {
     if (!mSocket.is_open()) {
       spdlog::error("Failed to write to the socket: not opened");
       return;
@@ -121,7 +97,7 @@ public:
 
     if constexpr (std::is_same_v<Socket, TcpSocket>) {
       boost::asio::async_write(
-          mSocket, boost::asio::buffer(dataPtr->data(), dataPtr->size()),
+          mSocket, boost::asio::buffer(dataPtr->data(), realSize),
           [this, dataPtr](BoostErrorRef ec, size_t size) { writeHandler(ec, size); });
     } else if constexpr (std::is_same_v<Socket, UdpSocket>) {
       mSocket.async_send_to(