--- a/common/src/network/async_socket.hpp
+++ b/common/src/network/async_socket.hpp
@@ -12,6 +12,7 @@
 #include <spdlog/spdlog.h>
 #include <vector>
 
+#include "constants.hpp"
 #include "network_types.hpp"
 #include "types.hpp"
 #include "utils/string_utils.hpp"