--- a/common/src/sink/command_sink.hpp
+++ b/common/src/sink/command_sink.hpp
@@ -9,6 +9,7 @@
 #include <functional>
 #include <map>
 
+#include "template_types.hpp"
 #include "utils/utils.hpp"
 
 namespace hft {