--- a/common/src/types/boost_types.hpp
+++ b/common/src/types/boost_types.hpp
@@ -7,8 +7,6 @@
 #define HFT_COMMON_BOOST_TYPES_HPP
 
 #include <boost/asio.hpp>
-#include <boost/fiber/all.hpp>
-#include <boost/lockfree/queue.hpp>
 
 namespace hft {
 
@@ -19,19 +17,11 @@ using ContextGuard = boost::asio::executor_work_guard<IoContext::executor_type>;
 using BoostError = boost::system::error_code;
 using BoostErrorRef = const BoostError &;
 
-using Fiber = boost::fibers::fiber;
-
 using SteadyTimer = boost::asio::steady_timer;
 using Seconds = boost::asio::chrono::seconds;
 using MilliSeconds = boost::asio::chrono::milliseconds;
 using Microseconds = boost::asio::chrono::microseconds;
 using Timestamp = std::chrono::time_point<std::chrono::system_clock>;
-
-template <typename EventType>
-using LFQueue = boost::lockfree::queue<EventType>;
-
-template <typename EventType>
-using UPtrLFQueue = std::unique_ptr<LFQueue<EventType>>;
 
 } // namespace hft