- Reworked IoSink making it buffer outgoing messages
- Reworked all event flow to work with std::span instead of single events
- Added EventFd class to improve busy waiting in sinks

Previously whenever match happens, response would be sent right away, and IoCtx loop gets congested with small messages. I added additional layer of lock free queues to IoSink. New event gets pushed into a queue and (if not yet) handler gets posted into the IoCtx that would process all the events at once. So buffering is only dependent on IoCtx loop load.

This shows some crazy improvements in RTT distribution, it completely eliminated all 1ms-50ms spikes

Before:
Server: [15:50:41.408] [I] [open|total]: 341401|535799 RPS:9220
Client:  [15:50:17.915] [I] RTT [1us|100us|1ms]  94.66% avg:24us  0.13% avg:433us  5.21% avg:25ms

After:
Server: [14:59:10.416] [I] [open|total]: 345135|541083 RPS:9026
Client:  [14:59:09.220] [I] RTT [1us|100us|1ms]  99.90% avg:26us  0.10% avg:132us  0.00% avg:0ms